+++ src/utils/reader.cpp
@@ -24,12 +24,12 @@ void Reader::readFirstProblem(const std::string problems_filepath){
   //for problem matrix) and convert to integer
   std::getline(newfile, temp_string);
   std::getline(newfile, num_rows_string);
-  const int num_rows = atoi(num_rows_string.c_str());
+  const int16_t num_rows = static_cast<int16_t>(atoi(num_rows_string.c_str()));
 
   //read in rows, add vectors to problem matrix
   for(size_t i = 0; i < num_rows; ++i){
     std::getline(newfile, temp_string);
-    std::vector<int> matrix_row = convertStringToVector(temp_string);
+    std::vector<int16_t> matrix_row = convertStringToVector(temp_string);
     problem_matrix_.push_back(matrix_row);
   }
 
@@ -37,14 +37,14 @@ void Reader::readFirstProblem(const std::string problems_filepath){
   std::string cost_vector_string;
   std::getline(newfile, cost_vector_string);
   std::getline(newfile, cost_vector_string);
-  std::vector<int> cost_vector = convertStringToVector(cost_vector_string);
+  std::vector<int16_t> cost_vector = convertStringToVector(cost_vector_string);
   for(size_t i = 0; i < cost_vector.size(); ++i){
     cost_vector_.push_back(cost_vector.at(i));
   }
   newfile.close();
 }
 
-void Reader::readProblem(const std::string problems_filepath, const int problem_number){
+void Reader::readProblem(const std::string problems_filepath, const int32_t problem_number){
   //clear anything still sored in class members from previous read
   problem_matrix_.clear();
   cost_vector_.clear();
@@ -57,7 +57,7 @@ void Reader::readProblem(const std::string problems_filepath, const int problem_
     std::cout << "ERROR: Unable to open file" << std::endl;
   }
   //find location in file we want to look at (nth problem)
-  int current_problem_location = 0;
+  int32_t current_problem_location = 0;
   std::string tempstring;
   std::string tilde = "~";
   while(current_problem_location != problem_number-1){
@@ -70,10 +70,10 @@ void Reader::readProblem(const std::string problems_filepath, const int problem_
   std::string numrows_string;
   std::getline(newfile, numrows_string);
   std::getline(newfile, numrows_string);
-  const int num_rows = atoi(numrows_string.c_str());
+  const int16_t num_rows = static_cast<int16_t>(atoi(numrows_string.c_str()));
 
   //get all rows, convert to integer vectors and add to matrix
-  std::vector<int> temp_vec;
+  std::vector<int16_t> temp_vec;
   for(size_t i = 0; i < num_rows; ++i){
     std::getline(newfile, tempstring);
     temp_vec = convertStringToVector(tempstring);
@@ -86,7 +86,7 @@ void Reader::readProblem(const std::string problems_filepath, const int problem_
   std::getline(newfile, cost_vector_string);
 
   //convert cost vector string to vector of ints and save as class member
-  const std::vector<int> cost_vector = convertStringToVector(cost_vector_string);
+  const std::vector<int16_t> cost_vector = convertStringToVector(cost_vector_string);
   for(size_t i = 0; i < cost_vector.size(); ++i){
     cost_vector_.push_back(cost_vector.at(i));
   }
@@ -95,7 +95,7 @@ void Reader::readProblem(const std::string problems_filepath, const int problem_
 }
 
 
-std::vector<int> Reader::convertStringToVector(const std::string vector_string){
+std::vector<int16_t> Reader::convertStringToVector(const std::string vector_string){
 
   std::string tempstring;
   std::vector<std::string> stringvec;
@@ -113,15 +113,15 @@ std::vector<int> Reader::convertStringToVector(const std::string vector_string){
   }
 
   //initialise variables
-  std::vector<int> rowvector;
-  int temp_int;
+  std::vector<int16_t> rowvector;
+  int16_t temp_int;
   std::string space = " ";
 
   //if string in vector is not empty or a space, convert to int and add to return vector
   for(size_t i = 0; i < stringvec.size(); ++i){
     tempstring = stringvec.at(i);
     if(tempstring.compare(space) != 0 && !tempstring.empty()){
-      temp_int = atoi(tempstring.c_str());
+      temp_int = static_cast<int16_t>(atoi(tempstring.c_str()));
       rowvector.push_back(temp_int);
     }
   }
